requirements.txt
aio-pika==9.5.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomysql==0.2.0
aiormq==6.8.1
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.4.2
loguru==0.7.2
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
omegaconf==2.3.0
pamqp==3.3.0
pika==1.3.2
prompt_toolkit==3.0.48
propcache==0.2.0
pycparser==2.22
pytest==9.0.2
pytest-asyncio==1.3.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
wechatpayv3==1.3.3
yarl==1.17.1
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.2
pycryptodome==3.20.0